@@ -1,43 +1,64 @@
 {
     "name": "@noelware/utils",
-    "description": "🌸 Extra utilities to keep code consist and non-repetitive.",
-    "version": "2.4.0",
-    "main": "dist/index.js",
-    "types": "dist/index.d.ts",
-    "author": "Noel Towa <cutie@floofy.dev>, Noelware Team <team@noelware.org>",
+    "description": "🐻‍❄️🌸 Extra utilities for Noelware's projects to keep code consistent",
+    "version": "2.5.0",
+    "type": "module",
+    "main": "./dist/index.cjs",
+    "module": "./dist/index.js",
+    "types": "./dist/index.d.ts",
+    "author": "Noel Towa <cutie@floofy.dev>",
     "license": "MIT",
     "repository": "https://github.com/Noelware/utils",
     "bugs": "https://github.com/Noelware/utils/issues",
     "homepage": "https://docs.noelware.org/libraries/@noelware/utils",
-    "packageManager": "yarn@3.8.0",
+    "exports": {
+        ".": {
+            "require": "./dist/index.cjs",
+            "import": "./dist/index.js",
+            "types": "./dist/index.d.ts"
+        },
+        "./result": {
+            "require": "./dist/result.cjs",
+            "import": "./dist/result.js",
+            "types": "./dist/result.d.ts"
+        },
+        "./option": {
+            "require": "./dist/option.cjs",
+            "import": "./dist/option.js",
+            "types": "./dist/option.d.ts"
+        }
+    },
     "files": [
-        "dist/"
+        "dist/",
+        "src/"
     ],
     "funding": {
         "url": "https://github.com/Noelware/utils?sponsor=1"
     },
     "engines": {
-        "node": ">=16"
+        "node": ">=18"
     },
     "scripts": {
-        "build:prod": "yarn lint && yarn fmt && node ./scripts/build.js",
-        "build": "yarn build:prod",
-        "lint": "eslint src --ext .ts --fix",
+        "build": "tsup",
+        "lint": "bun scripts/eslint.ts",
         "test": "vitest --run",
-        "fmt": "prettier --write ./.prettierrc.json **/*.{ts,js,yaml,yml,json,md}"
+        "fmt": "bun scripts/prettier.ts"
     },
     "devDependencies": {
+        "@actions/core": "1.10.1",
         "@augu/eslint-config": "5.1.1",
         "@augu/tsconfig": "1.2.0",
+        "@types/bun": "1.0.5",
         "@types/node": "20.11.17",
         "@typescript-eslint/eslint-plugin": "6.21.0",
         "@typescript-eslint/parser": "6.21.0",
+        "colorette": "2.0.20",
         "eslint": "8.45.0",
         "eslint-config-prettier": "9.1.0",
+        "eslint-formatter-codeframe": "7.32.1",
+        "eslint-plugin-perfectionist": "2.5.0",
         "prettier": "3.2.5",
         "tsup": "8.0.2",
-        "typescript": "5.3.3",
-        "vite": "5.1.1",
-        "vitest": "0.33.0"
+        "typescript": "5.3.3"
     }
 }
